Share Love Tab Created
Added in the share love tab and icon. Page has a functional number pad and display along with backspace. When the user selects to send their payment, it will display a message that will tell them the payment was sent and how much was sent. The number entered label will then be cleared.